Moving your user class into a different file and cleaning up the way argparse is used
